+++ I18n/_i18n/Controller.ts
@@ -5,11 +5,13 @@ import Store from './Store';
 import Translator from './Translator';
 
 import IContext from './interfaces/IContext';
-import ILoader, {ILoadingsHistory} from './interfaces/ILoader';
+import ILoader from './interfaces/ILoader';
 import IController from './interfaces/IController';
-import ILocale from '../locales/Interfaces/ILocale';
-import {IModule} from './interfaces/declaration';
+import ILocale from '../locales/interfaces/ILocale';
 import ITranslator from './interfaces/ITranslator';
+import IStore from './interfaces/IStore';
+import ILoadingsHistory from './interfaces/ILoadingsHistory';
+import IModule from './interfaces/IModule';
 
 const LENGTH_LANG_CODE = 2;
 const LENGTH_LOCALE_CODE = 5;
@@ -31,21 +33,39 @@ interface IRequest {
     };
 }
 
+/**
+ * Контролер для работы и взаимодействия с механизмом локализации.
+ * @class I18n/_i18n/Controller
+ * @implements Ii18n/_i18n/interfaces/IController
+ * @public
+ * @author Кудрявцев И.С.
+ */
 class Controller implements IController {
     contextSeparator: string = '@@';
     pluralPrefix: string = 'plural#';
     pluralDelimiter: string = '|';
 
+    /**
+     * Дефолтная локаль приложения.
+     * @type String
+     */
+    readonly defaultLocale: string = '';
+
+    /**
+     * Загрузчик ресурсов локализации.
+     * @type I18n/_i18n/interfaces/ILoader
+     * @default {@link I18n/_i18n/Loader Loader}
+     */
+    readonly loader: ILoader;
+
     protected availableLocales: string[] = [];
     protected availableContexts: {[contextName: string]: IModule} = {};
-    readonly defaultLocale: string = '';
     protected defaultLocalesFromLang: { [lang: string]: string } = {};
     protected currentCodeLocale: string = '';
     protected loadableTranslator: { [context: string]: Promise<ITranslator> } = {};
-    private translators: { [context: string]: ITranslator } = {};
-    private localesStore: Store<ILocale>;
-    private contextStore: Store<IContext>;
-    readonly loader: ILoader;
+    protected translators: { [context: string]: ITranslator } = {};
+    protected localesStore: IStore<ILocale>;
+    protected contextStore: IStore<IContext>;
 
     constructor(config: IConfigController) {
         this.readConfig(config);
@@ -57,10 +77,40 @@ class Controller implements IController {
         this.initStores();
     }
 
+    protected get requiredLocales(): string[] {
+        if (!this.isEnabled) {
+            return [this.defaultLocale];
+        }
+
+        return constants.isServerSide ? this.availableLocales : [this.currentLocale];
+    }
+
+    /**
+     * Дефолтный язык приложения.
+     * @return {String}
+     */
     get defaultLang(): string {
         return this.defaultLocale.split('-')[0];
     }
 
+    /**
+     * Код установленной локали приложения.
+     * Если не удалось определить код локали или выключенал локализация, вернёт дефолтную локаль.
+     * @example
+     * Приложение отображается в англо-амереканской локале.
+     * <pre>
+     *    controller.currentLocale === 'ru-RU' // false
+     *    controller.currentLocale === 'en-US' // true
+     * </pre>
+     *
+     * @example
+     * Локализауия выключена.
+     * <pre>
+     *    controller.currentLocale === 'ru-RU' // true
+     * </pre>
+     *
+     * @return {String}
+     */
     get currentLocale(): string {
         if (!this.isEnabled) {
             return this.defaultLocale;
@@ -77,30 +127,58 @@ class Controller implements IController {
         }
     }
 
+    /**
+     * Установленный язык приложения.
+     * Если не удалось определить код языка или выключенал локализация, вернёт дефолтный язык.
+     * @example
+     * Приложение отображается в англо-амереканской локале.
+     * <pre>
+     *    controller.currentLang === 'ru' // false
+     *    controller.currentLang === 'en-US' // false
+     *    controller.currentLang === 'en' // true
+     * </pre>
+     *
+     * @example
+     * Локализауия выключена.
+     * <pre>
+     *    controller.currentLang === 'ru' // true
+     * </pre>
+     * @return {String}
+     */
     get currentLang(): string {
         return this.currentLocale.split('-')[0];
     }
 
+    /**
+     * История загрузки ресурсов локализации.
+     * @return {I18n/_i18n/interfaces/ILoader:ILoadingsHistory}
+     */
     get loadingsHistory(): ILoadingsHistory {
         return this.loader.history;
     }
 
-    get requiredLocales(): string[] {
-        if (!this.isEnabled) {
-            return [this.defaultLocale];
-        }
-
-        return constants.isServerSide ? this.availableLocales : [this.currentLocale];
-    }
-
+    /**
+     * Конфигурация установленной локали прилолжения.
+     * @return {I18n/locales/interfaces/ILocale}
+     */
     get currentLocaleConfig(): ILocale {
         return this.localesStore.get(this.currentLocale, true) as ILocale || {} as ILocale;
     }
 
+    /**
+     * Включена ли локализация.
+     * @return {Boolean}
+     */
     get isEnabled(): boolean {
         return this.availableLocales.length !== 0;
     }
 
+    /**
+     * Устанавливает код локали в куку lang.
+     * @param {String} code Код локали.
+     * @param {Number} [expires=365] Время жизни куки.
+     * @return {Void}
+     */
     setLocale(code: string, expires: number = EXPIRES_COOKIES): void {
         const fullCode = this._normalizeCode(code);
 
@@ -114,10 +192,19 @@ class Controller implements IController {
         }
     }
 
+    /**
+     * Проверяет поддерижвается ли код локали или языка.
+     * @param {String} code Код локали.
+     * @return {Boolean}
+     */
     isSupportedLocale(code: string): boolean {
         return this._isSupportedLocale(this._normalizeCode(code));
     }
 
+    /**
+     * Сигнализирует о готовности контролера.
+     * @return {Promise<Boolean>}
+     */
     isReady(): Promise<boolean> {
         return new Promise((resolve, reject) => {
             const locales = [];
@@ -132,6 +219,12 @@ class Controller implements IController {
         });
     }
 
+    /**
+     * Возврашает переводчик для контекста.
+     * @param {String} contextName Имя контекста.
+     * @param {Boolean} [sync=false] Получить переводчик синхроно, если это первый запрос за ним, вернёт undefined.
+     * @return {Promise<I18n/_i18n/interfaces/ITranslator> | I18n/_i18n/interfaces/ITranslator}
+     */
     getTranslator(contextName: string, sync: boolean = false): Promise<ITranslator> | ITranslator {
         if (this.translators.hasOwnProperty(contextName)) {
             delete this.loadableTranslator[contextName];
@@ -166,10 +259,23 @@ class Controller implements IController {
         return this.loadableTranslator[contextName];
     }
 
+    /**
+     * Добавляет контекст в хранилище.
+     * @param {String} contextName Имя контекста.
+     * @param {I18n/_i18n/interfaces/IContext} [context] Ресусры контекста.
+     * @return {Void}
+     */
     addContext(contextName: string, context?: IContext): void {
         this.contextStore.set(contextName, context);
     }
 
+    /**
+     * Добавляет локаль в хранилище.
+     * @param {String} localeCode Код локали.
+     * @param {I18n/locales/interfaces/ILocale} [locale] конфигурация локали.
+     * @param [isAvailable=true] Надо ли добавить локаль в доступные.
+     * @return {Void}
+     */
     addLocale(localeCode: string, locale?: ILocale, isAvailable: boolean = true): void {
         if (isAvailable) {
             this.availableLocales.push(localeCode);
@@ -179,7 +285,7 @@ class Controller implements IController {
         this.localesStore.set(localeCode, locale);
     }
 
-    protected readConfig(config: IConfigController): void {
+    private readConfig(config: IConfigController): void {
         for (const nameOption of Object.keys(config)) {
             if (config[nameOption] !== undefined) {
                 this[nameOption] = config[nameOption];
@@ -187,7 +293,7 @@ class Controller implements IController {
         }
     }
 
-    protected initStores(): void {
+    private initStores(): void {
         this.localesStore = new Store<ILocale>(
             this.requiredLocales,
             (localeCode: string) => this.loader.locale(localeCode)
@@ -198,7 +304,7 @@ class Controller implements IController {
         );
     }
 
-    protected buildMapOfDefaultLocales(): void {
+    private buildMapOfDefaultLocales(): void {
         for (const code of this.availableLocales) {
             const codeLang = code.split('-')[0];
 
@@ -287,6 +393,16 @@ class Controller implements IController {
         return '';
     }
 
+    /**
+     * Возврашает языки из заголовка accept-language.
+     * @param {IRequest} [request]  Тело https запроса.
+     * @example
+     * В accept-language лежит en-US,en;q=0.9,ru;q=0.8.
+     * <pre>
+     *    Controller.getAcceptLanguage() // ['en-US', 'en;q=0.9', 'ru;q=0.8']
+     * </pre>
+     * @return {String[]}
+     */
     static getAcceptLanguage(request?: IRequest): string[] {
         // @ts-ignore
         const req = request || process && process.domain && process.domain.req;
@@ -295,10 +411,32 @@ class Controller implements IController {
         return acceptLanguage ? acceptLanguage.split(',') : [];
     }
 
+    /**
+     * Проверяет явлется ли строка кодом языка.
+     * @param {String} code Проверяемы код.
+     * @example
+     * <pre>
+     *    Controller.isLangCode('en-US') // false
+     *    Controller.isLangCode('en') // true
+     *    Controller.isLangCode('gg') // true
+     * </pre>
+     * @return {Boolean}
+     */
     static isLangCode(code: string): boolean {
         return code.length === LENGTH_LANG_CODE;
     }
 
+    /**
+     * Проверяет явлется ли строка кодом локали.
+     * @param {String} code Проверяемы код.
+     * @example
+     * <pre>
+     *    Controller.isLocaleCode('en') // false
+     *    Controller.isLocaleCode('en-US') // true
+     *    Controller.isLocaleCode('gg-gg') // true
+     * </pre>
+     * @return {Boolean}
+     */
     static isLocaleCode(code: string): boolean {
         return code.length === LENGTH_LOCALE_CODE;
     }
