+++ I18n/_i18n/interfaces/IContents.ts
@@ -1,0 +1,16 @@
+import IModule from './IModule';
+
+/**
+ * Интерфейс объектов, содержащие информацию о доступных ресурсах локализации для приложения.
+ * @interface I18n/_i18n/interfaces/IContents
+ * @public
+ * @author Кудрявцев И.С.
+ */
+export default interface IContents {
+    buildnumber: string;
+    modules: {
+        [key: string]: IModule;
+    };
+    availableLanguage?: object;
+    defaultLanguage?: string;
+}
