+++ I18n/_i18n/interfaces/IContext.ts
@@ -1,7 +1,11 @@
-export interface IDictionary {
-    [key: string]: string;
-}
+import IDictionary from './IDictionary';
 
+/**
+ * Интерфейс объектов, содержащие ресурсы для контеста локализации.
+ * @interface I18n/_i18n/interfaces/IContext
+ * @public
+ * @author Кудрявцев И.С.
+ */
 export default interface IContext {
     [localeCode: string]: IDictionary;
 }
