+++ I18n/_i18n/interfaces/IDictionary.ts
@@ -1,0 +1,9 @@
+/**
+ * Интерфейс словаря локализации.
+ * @interface I18n/_i18n/interfaces/IDictionary
+ * @public
+ * @author Кудрявцев И.С.
+ */
+export default interface IDictionary {
+    [key: string]: string;
+}
