--- I18n/_i18n/keyboard.ts
@@ -1,1 +1,0 @@
-export {default as changeLayout} from './keyboard/changeLayout';
