+++ I18n/i18n.ts
@@ -1,17 +1,15 @@
 /**
- * Library for internationalization
+ * Библиотека для работы с локализацией.
  * @library I18n/i18n
- * @includes Locale I18n/_i18n/Locale
  * @includes Translator I18n/_i18n/Translator
- * @includes controller I18n/_i18n/Controller
- * @includes Configuration I18n/_i18n/Configuration
  * @includes Loader I18n/_i18n/Loader
- * @includes keyboard I18n/_i18n/keyboard
+ * @includes controller I18n/_i18n/Controller
+ * @includes keyboard I18n/keyboard
  * @public
  * @author Кудрявцев И.С.
  */
 
-import * as keyboard from './_i18n/keyboard';
+import * as keyboard from './keyboard';
 import controller, {load} from './_i18n/SingletonController';
 
 export {
@@ -19,6 +17,7 @@ export {
     load,
     controller
 };
+
 export {default as Translator} from './_i18n/Translator';
 export {default as Controller} from './_i18n/Controller';
 export {default as Locale} from './_i18n/Locale';
