+++ I18n/keyboard.ts
@@ -1,0 +1,9 @@
+/**
+ * Набор методов для работы с клавиатурной раскалдкой.
+ * @library I18n/keyboard
+ * @includes changeLayout I18n/_keyboard/changeLayout
+ * @public
+ * @author Кудрявцев И.С.
+ */
+
+export {default as changeLayout} from './_keyboard/changeLayout';
