--- I18n/locales/Interfaces/INumber.ts
@@ -1,4 +1,0 @@
-export default interface INumber {
-    fractionSeparator: string;
-    triadDelimiter: string;
-}
