+++ I18n/locales/_aspects/Number/USA.ts
@@ -1,4 +1,4 @@
-import INumber from '../../Interfaces/INumber';
+import INumber from '../../interfaces/INumber';
 
 export default class USA implements INumber {
     fractionSeparator: string = '.';
