+++ I18n/locales/en-US.ts
@@ -8,11 +8,18 @@ import NumberAspect from './_aspects/Number/Europe';
 import CalendarEntitiesAspect from './_aspects/CalendarEntities/English';
 import pluralFunction from './_aspects/Plural/English';
 
-import ILocale from './Interfaces/ILocale';
-import IDate from './Interfaces/IDate';
-import INumber from './Interfaces/INumber';
-import ICalendarEntities from './Interfaces/ICalendarEntities';
+import ILocale from './interfaces/ILocale';
+import IDate from './interfaces/IDate';
+import INumber from './interfaces/INumber';
+import ICalendarEntities from './interfaces/ICalendarEntities';
 
+/**
+ * Конфигурация для англо-амереканcкой локали.
+ * @class I18n/locales/en-US
+ * @implements I18n/interfaces/ILocale
+ * @public
+ * @author Кудрявцев И.С.
+ */
 class ENUS implements ILocale {
     code: string = 'en-US';
     date: IDate = new DateAspect();
