+++ I18n/locales/interfaces/ICalendarEntities.ts
@@ -1,3 +1,9 @@
+/**
+ * Интерфейс конфигурации для локализация календарнной обозначений(Дни недели, месяц, квартал).
+ * @interface I18n/locales/interfaces/ICalendarEntities
+ * @public
+ * @author Кудрявцев И.С.
+ */
 export default interface ICalendarEntities {
    minDays: string[];
    shortDays: string[];
