+++ I18n/locales/interfaces/IDate.ts
@@ -1,3 +1,9 @@
+/**
+ * Интерфейс конфигурации для локализация дат.
+ * @interface I18n/locales/interfaces/IDate
+ * @public
+ * @author Кудрявцев И.С.
+ */
 export default interface IDate {
     fullDateFormat: string;
     fullDateFullYearFormat: string;
