+++ I18n/locales/interfaces/ILocale.ts
@@ -2,6 +2,12 @@ import INumber from './INumber';
 import ICalendarEntities from './ICalendarEntities';
 import IDate from './IDate';
 
+/**
+ * Интерфейс конфигурации локали.
+ * @interface I18n/locales/interfaces/ILocale
+ * @public
+ * @author Кудрявцев И.С.
+ */
 export default interface ILocale {
     code: string;
     date: IDate;
