+++ I18n/locales/interfaces/INumber.ts
@@ -1,0 +1,10 @@
+/**
+ * Интерфейс конфигурации для локализация числовых значений.
+ * @interface I18n/locales/interfaces/INumber
+ * @public
+ * @author Кудрявцев И.С.
+ */
+export default interface INumber {
+    fractionSeparator: string;
+    triadDelimiter: string;
+}
