+++ I18n/locales/ru-RU.ts
@@ -3,10 +3,10 @@ import NumberAspect from './_aspects/Number/Europe';
 import CalendarEntitiesAspect from './_aspects/CalendarEntities/Russian';
 import pluralFunction from './_aspects/Plural/Russian';
 
-import ILocale from './Interfaces/ILocale';
-import IDate from './Interfaces/IDate';
-import INumber from './Interfaces/INumber';
-import ICalendarEntities from './Interfaces/ICalendarEntities';
+import ILocale from './interfaces/ILocale';
+import IDate from './interfaces/IDate';
+import INumber from './interfaces/INumber';
+import ICalendarEntities from './interfaces/ICalendarEntities';
 
 class EuropeRuDate extends DateAspect {
     shortQuarterFormat: string = 'QQr \'YY';
@@ -21,6 +21,13 @@ class EuropeRuDate extends DateAspect {
     shortDateShortMonthFormat: string = 'DD MMMl';
 }
 
+/**
+ * Конфигурация для русской локали.
+ * @class I18n/locales/ru-RU
+ * @implements I18n/interfaces/ILocale
+ * @public
+ * @author Кудрявцев И.С.
+ */
 class RURU implements ILocale {
     code: string = 'ru-RU';
     date: IDate = new EuropeRuDate();
